refactor(xapiri): extract dashboard_snapshot.go — buildSnapshotCfg, flushToCfg, markDirty
Part 4 of ADR 0014 dashboard split. Moves the paired atomic state
serialisers (buildSnapshotCfg + flushToCfg) and markDirty to a
dedicated file. Any diff touching state persistence is now scoped
to dashboard_snapshot.go.
